CaseInsensitiveLogin/requirements.txt
Flask
flask-login 
flask-sqlalchemy
email-validator
python-dotenv